refactor: migrate feature-specific files from packages/lib to packages/features
- Move bookings-related files to packages/features/bookings/lib
- Move schedules-related files to packages/features/schedules/repositories
- Move repository files to appropriate feature directories
- Update all import paths across the codebase
- Fix relative imports in moved files
- Fix lint warnings in getBookingFields.ts and HostEditDialogs.tsx
- Fix unused expression warnings in CalEventParser.ts
