+++ apps/web/app/cache/travelSchedule.ts
@@ -2,8 +2,8 @@
 
 import { revalidateTag } from "next/cache";
 
+import { TravelScheduleRepository } from "@calcom/features/schedules/repositories/travelSchedule";
 import { NEXTJS_CACHE_TTL } from "@calcom/lib/constants";
-import { TravelScheduleRepository } from "@calcom/lib/server/repository/travelSchedule";
 import { unstable_cache } from "@calcom/lib/unstable_cache";
 
 const CACHE_TAGS = {
