+++ apps/web/modules/test-setup.ts
@@ -60,7 +60,7 @@ vi.mock("@calcom/features/bookings/components/event-meta/Price", () => {
   return {};
 });
 
-vi.mock("@calcom/lib/bookings/SystemField", () => {
+vi.mock("@calcom/features/bookings/lib/SystemField", () => {
   return {};
 });
