+++ packages/features/bookings/lib/hostGroupUtils.ts
