+++ packages/features/bookings/lib/routing/utils.test.ts
