+++ packages/features/bookings/lib/routing/utils.ts
