+++ packages/features/di/modules/Schedule.ts
@@ -1,5 +1,5 @@
 import { DI_TOKENS } from "@calcom/features/di/tokens";
-import { ScheduleRepository } from "@calcom/lib/server/repository/schedule";
+import { ScheduleRepository } from "@calcom/features/schedules/repositories/schedule";
 
 import { createModule } from "../di";
