+++ packages/features/eventtypes/components/CheckedTeamSelect.tsx
@@ -5,8 +5,8 @@ import { useState } from "react";
 import type { Options, Props } from "react-select";
 
 import { useIsPlatform } from "@calcom/atoms/hooks/useIsPlatform";
+import { getHostsFromOtherGroups } from "@calcom/features/bookings/lib/hostGroupUtils";
 import type { SelectClassNames } from "@calcom/features/eventtypes/lib/types";
-import { getHostsFromOtherGroups } from "@calcom/lib/bookings/hostGroupUtils";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import classNames from "@calcom/ui/classNames";
 import { Avatar } from "@calcom/ui/components/avatar";
