+++ packages/features/schedules/lib/transformers/for-atom.ts
@@ -52,7 +52,7 @@ export function transformDateOverridesForAtom(
     };
     const dayRangeIndex = acc.findIndex(
       // early return prevents override.date from ever being empty.
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+       
       (item) => yyyymmdd(item.ranges[0].start) === yyyymmdd(override.date!)
     );
     if (dayRangeIndex === -1) {
