+++ packages/features/schedules/lib/transformers/getScheduleListItemData.ts
