+++ packages/platform/atoms/availability/wrappers/AvailabilitySettingsPlatformWrapper.tsx
@@ -2,7 +2,7 @@ import type { ReactNode } from "react";
 import { forwardRef, useRef } from "react";
 
 import type { ScheduleLabelsType } from "@calcom/features/schedules/components/Schedule";
-import type { UpdateScheduleResponse } from "@calcom/lib/schedules/updateSchedule";
+import type { UpdateScheduleResponse } from "@calcom/features/schedules/lib/updateSchedule";
 import type { ApiErrorResponse, ApiResponse, UpdateScheduleInput_2024_06_11 } from "@calcom/platform-types";
 
 import { useAtomSchedule } from "../../hooks/schedules/useAtomSchedule";
