+++ packages/platform/atoms/hooks/schedules/useAtomSchedule.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 
-import type { FindDetailedScheduleByIdReturnType } from "@calcom/lib/server/repository/schedule";
+import type { FindDetailedScheduleByIdReturnType } from "@calcom/features/schedules/repositories/schedule";
 import { SUCCESS_STATUS } from "@calcom/platform-constants";
 import type { ApiResponse } from "@calcom/platform-types";
