+++ packages/platform/atoms/hooks/schedules/useEnsureDefaultSchedule.ts
@@ -1,6 +1,6 @@
 import { useEffect, useRef } from "react";
 
-import { Schedule } from "@calcom/lib/schedules/transformers/getTransformedSchedles";
+import { Schedule } from "@calcom/features/schedules/lib/transformers/getTransformedSchedles";
 
 export function useEnsureDefaultSchedule(
   schedules: Schedule[],
