+++ packages/platform/libraries/repositories.ts
@@ -1,10 +1,10 @@
 export { PrismaAttributeRepository } from "@calcom/lib/server/repository/PrismaAttributeRepository";
 export { PrismaOOORepository } from "@calcom/lib/server/repository/ooo";
-export { ScheduleRepository as PrismaScheduleRepository } from "@calcom/lib/server/repository/schedule";
+export { ScheduleRepository as PrismaScheduleRepository } from "@calcom/features/schedules/repositories/schedule";
 export { BookingRepository as PrismaBookingRepository } from "@calcom/features/bookings/repositories/BookingRepository";
 export { EventTypeRepository as PrismaEventTypeRepository } from "@calcom/features/eventtypes/repositories/eventTypeRepository";
 export { RoutingFormResponseRepository as PrismaRoutingFormResponseRepository } from "@calcom/lib/server/repository/formResponse";
-export { PrismaSelectedSlotRepository } from "@calcom/lib/server/repository/PrismaSelectedSlotRepository";
+export { PrismaSelectedSlotRepository } from "@calcom/features/bookings/repositories/PrismaSelectedSlotRepository";
 export { TeamRepository as PrismaTeamRepository } from "@calcom/features/ee/teams/repositories/TeamRepository";
 export { UserRepository as PrismaUserRepository } from "@calcom/features/users/repositories/UserRepository";
 export { FeaturesRepository as PrismaFeaturesRepository } from "@calcom/features/flags/features.repository";
