+++ packages/trpc/server/routers/viewer/availability/schedule/get.handler.ts
@@ -1,4 +1,4 @@
-import { ScheduleRepository } from "@calcom/lib/server/repository/schedule";
+import { ScheduleRepository } from "@calcom/features/schedules/repositories/schedule";
 
 import type { TrpcSessionUser } from "../../../../types";
 import type { TGetInputSchema } from "./get.schema";
