+++ packages/trpc/server/routers/viewer/slots/reserveSlot.handler.test.ts
@@ -12,7 +12,7 @@ const dynamicImportHandler = async () => await import("./reserveSlot.handler");
 
 // The repository instance method is used to check for an existing reservation by someone else.
 // To keep this unit test isolated from the database layer, we stub this to always resolve falsey.
-vi.mock("@calcom/lib/server/repository/PrismaSelectedSlotRepository", () => ({
+vi.mock("@calcom/features/bookings/repositories/PrismaSelectedSlotRepository", () => ({
   PrismaSelectedSlotRepository: vi.fn().mockImplementation(() => ({
     findReservedByOthers: vi.fn().mockResolvedValue(null),
   })),
