+++ packages/trpc/server/routers/viewer/slots/reserveSlot.handler.ts
@@ -3,9 +3,9 @@ import type { NextApiRequest, NextApiResponse } from "next";
 import { v4 as uuid } from "uuid";
 
 import dayjs from "@calcom/dayjs";
+import { PrismaSelectedSlotRepository } from "@calcom/features/bookings/repositories/PrismaSelectedSlotRepository";
 import { WEBAPP_URL } from "@calcom/lib/constants";
 import { MINUTES_TO_BOOK } from "@calcom/lib/constants";
-import { PrismaSelectedSlotRepository } from "@calcom/lib/server/repository/PrismaSelectedSlotRepository";
 import type { PrismaClient } from "@calcom/prisma";
 import { BookingStatus } from "@calcom/prisma/enums";
