+++ packages/trpc/server/routers/viewer/travelSchedules/getTravelSchedules.handler.ts
@@ -1,4 +1,4 @@
-import { TravelScheduleRepository } from "@calcom/lib/server/repository/travelSchedule";
+import { TravelScheduleRepository } from "@calcom/features/schedules/repositories/travelSchedule";
 import type { TrpcSessionUser } from "@calcom/trpc/server/types";
 
 type GetTravelSchedulesOptions = {
